Move code to new object from JSONEditor (#89)
